requirements.txt
anyio==3.2.1
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
Babel==2.9.1
backcall==0.2.0
bleach==3.3.0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
Cython==0.29.23
decorator==5.0.9
defusedxml==0.7.1
Deprecated==1.2.12
dnspython==2.0.0
entrypoints==0.3
filelock==3.0.12
idna==2.10
ipwhois==1.2.0
ipykernel==5.5.5
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-server==1.9.0
jupyterlab==3.0.16
jupyterlab-pygments==0.1.2
jupyterlab-server==2.6.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclassic==0.3.1
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
neo4j==4.3.1
nest-asyncio==1.5.1
notebook==6.4.0
numpy==1.21.0
packaging==20.9
pandas==1.3.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prometheus-client==0.11.0
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
pymisp==2.4.144
pyparsing==2.4.7
pypdns==1.5.2
pyrsistent==0.18.0
python-dateutil==2.8.1
python-dotenv==0.18.0
pytz==2021.1
pyzmq==22.1.0
requests==2.25.1
requests-cache==0.6.4
requests-file==1.5.1
requests-unixsocket==0.2.0
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.7.1
six==1.16.0
sniffio==1.2.0
sparse-dot-topn==0.3.1
string-grouper==0.5.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.1.0
tldextract==3.1.0
tornado==6.1
traitlets==5.0.5
url-normalize==1.4.3
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.1.0
wrapt==1.12.1
